@@ -1,9 +1,8 @@
 group 'uk.mollies.lunchmancore'
-version '1.0-SNAPSHOT'
 
 apply plugin: 'java'
 
-sourceCompatibility = 1.5
+sourceCompatibility = 1.8
 
 repositories {
     mavenCentral()
