@@ -2,7 +2,7 @@ package LunchManCore;
 
 import org.junit.Test;
 
-import java.util.Date;
+import java.time.LocalDate;
 
 import static org.junit.Assert.assertEquals;
 
@@ -10,7 +10,16 @@ public class FridayLunchTest {
 
     @Test
     public void hasADate() {
-        FridayLunch fridayLunch = new FridayLunch(new Date(20160611));
-        assertEquals(new Date(20160611), fridayLunch.getDate());
+        FridayLunch fridayLunch = new FridayLunch(LocalDate.of(2016, 6, 10));
+        assertEquals(LocalDate.of(2016, 6, 10), fridayLunch.getDate());
     }
+
+    @Test
+    public void hasAnApprentice() {
+        FridayLunch fridayLunch = new FridayLunch(LocalDate.of(2016, 6, 10));
+        Apprentice nick = new Apprentice("Nick");
+        fridayLunch.assignApprentice(nick);
+        assertEquals(nick, fridayLunch.getApprentice().get());
+    }
+
 }
